New features and attempted bug fixes
Working on adding bombs as a weapon against zombies and trying to figure out bugs. There are even more bugs now, but at least they're more noticeable so hopefully easier to find/fix. Eventually I hope to create achievements and other types of weapons/lures, but the bomb thing and bug fixes are all I'm focused on for now.